util: move runningPlatform() to nicstobridge.go
There's only one user so put the function there.
