+++ go-controller/pkg/util/ovs_unit_test.go
@@ -30,69 +30,6 @@ type onCallReturnArgs struct {
 	retArgList          []interface{}
 }
 
-func TestRunningPlatform(t *testing.T) {
-	// Below is defined in ovs.go file
-	AppFs = afero.NewMemMapFs()
-	AppFs.MkdirAll("/etc", 0755)
-	tests := []struct {
-		desc            string
-		fileContent     []byte
-		filePermissions os.FileMode
-		expOut          string
-		expErr          error
-	}{
-		{
-			desc:   "ReadFile returns error",
-			expErr: fmt.Errorf("failed to parse file"),
-		},
-		{
-			desc:            "failed to find platform name",
-			expErr:          fmt.Errorf("failed to find the platform name"),
-			fileContent:     []byte("NAME="),
-			filePermissions: 0755,
-		},
-		{
-			desc:            "platform name returned is RHEL",
-			expOut:          "RHEL",
-			fileContent:     []byte("NAME=\"CentOS Linux\""),
-			filePermissions: 0755,
-		},
-		{
-			desc:            "platform name returned is Ubuntu",
-			expOut:          "Ubuntu",
-			fileContent:     []byte("NAME=\"Debian\""),
-			filePermissions: 0755,
-		},
-		{
-			desc:            "platform name returned is Photon",
-			expOut:          "Photon",
-			fileContent:     []byte("NAME=\"VMware\""),
-			filePermissions: 0755,
-		},
-		{
-			desc:            "unknown platform",
-			expErr:          fmt.Errorf("unknown platform"),
-			fileContent:     []byte("NAME=\"blah\""),
-			filePermissions: 0755,
-		},
-	}
-	for i, tc := range tests {
-		t.Run(fmt.Sprintf("%d:%s", i, tc.desc), func(t *testing.T) {
-			if tc.fileContent != nil && tc.filePermissions != 0 {
-				afero.WriteFile(AppFs, "/etc/os-release", tc.fileContent, tc.filePermissions)
-				defer AppFs.Remove("/etc/os-release")
-			}
-			res, err := runningPlatform()
-			t.Log(res, err)
-			if tc.expErr != nil {
-				assert.Contains(t, err.Error(), tc.expErr.Error())
-			} else {
-				assert.Equal(t, res, tc.expOut)
-			}
-		})
-	}
-}
-
 func TestRunOVNretry(t *testing.T) {
 	mockKexecIface := new(mock_k8s_io_utils_exec.Interface)
 	mockExecRunner := new(mocks.ExecRunner)
